requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==4.2.13
filelock==3.14.0
flask==3.0.3
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
platformdirs==4.2.2
psycopg2==2.9.3
PyJWT==2.8.0
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
virtualenv==20.26.2
werkzeug==3.0.3
zipp==3.18.2
